using __init__ as is supposed to
